refactor: MoveDao를 바로 클래스로 사용하도록 변경
